feat: add WelcomeScreen and LogDisplay components for improved navigation and log viewing
