+++ app/src/main/java/com/aoscoremonitor/MainActivity.kt
@@ -4,23 +4,18 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import com.aoscoremonitor.diagnostics.LogCollector
-import com.aoscoremonitor.ui.screen.LogDisplay
+import com.aoscoremonitor.ui.screens.LogDisplay
+import com.aoscoremonitor.ui.screens.WelcomeScreen
 import com.aoscoremonitor.ui.theme.AOSCoreMonitorTheme
 
 class MainActivity : ComponentActivity() {
@@ -66,20 +61,5 @@ class MainActivity : ComponentActivity() {
     override fun onPause() {
         super.onPause()
         logCollector.stopCollecting()
-    }
-
-    @Composable
-    fun WelcomeScreen(onNavigateToLogs: () -> Unit, modifier: Modifier = Modifier) {
-        Box(
-            modifier = modifier.fillMaxSize(),
-            contentAlignment = Alignment.Center
-        ) {
-            Button(
-                onClick = onNavigateToLogs,
-                modifier = Modifier.padding(16.dp)
-            ) {
-                Text(text = "Hello Android!")
-            }
-        }
     }
 }
