+++ app/src/main/java/com/aoscoremonitor/ui/screens/LogScreen.kt
@@ -1,4 +1,4 @@
-package com.aoscoremonitor.ui.screen
+package com.aoscoremonitor.ui.screens
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
