+++ app/src/main/java/com/aoscoremonitor/ui/screens/WelcomeScreen.kt
@@ -1,0 +1,26 @@
+package com.aoscoremonitor.ui.screens
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun WelcomeScreen(onNavigateToLogs: () -> Unit, modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Button(
+            onClick = onNavigateToLogs,
+            modifier = Modifier.padding(16.dp)
+        ) {
+            Text(text = "Hello Android!")
+        }
+    }
+}
